cleanup and de-dupe LandingHero implementation
