+++ src/components/landing-hero/types.ts
@@ -1,6 +1,6 @@
 import { ProductRootDocsPathLandingProps } from 'views/product-root-docs-path-landing/types'
 
-export interface ProductRootDocsPathLandingHeroProps {
+export interface LandingHeroProps {
 	pageHeading: ProductRootDocsPathLandingProps['pageHeading']
 	pageSubtitle: string
 }
