+++ src/views/docs-view/components/docs-root-landing/index.tsx
@@ -1,5 +1,5 @@
-import { Children, useMemo } from 'react'
-import LandingHero from '../hero'
+import { Children, ReactChild } from 'react'
+import LandingHero from 'components/landing-hero'
 
 function getElementTextContent(child) {
 	let textContent = ''
@@ -13,19 +13,34 @@ function getElementTextContent(child) {
 	return textContent
 }
 
-export default function DocsRootLanding({ children, heading, subtitle }) {
-	let firstH1
-	let firstP
+/**
+ * TEMPORARY
+ *
+ * A structured page view that wraps rendered MDX content, extracts the first rendered H1 and p tags and renders a
+ * landing hero in their place. Used to demonstrate structured page views pattern driven by frontmatter
+ */
+export default function DocsRootLanding({
+	children,
+}: {
+	children: ReactChild[]
+}) {
+	let heading
+	let subtitle
+
 	const childrenToRender = Children.map(children, (child) => {
-		if (!firstH1 && child.props.originalType === 'h1') {
-			firstH1 = {
+		if (typeof child === 'string' || typeof child === 'number') {
+			return child
+		}
+
+		if (!heading && child.props.originalType === 'h1') {
+			heading = {
 				title: getElementTextContent(child),
 				id: child.props.id,
 			}
 			return null
 		}
-		if (!firstP && child.props.originalType === 'p') {
-			firstP = child
+		if (!subtitle && child.props.originalType === 'p') {
+			subtitle = child
 			return null
 		}
 		return child
@@ -33,7 +48,7 @@ export default function DocsRootLanding({ children, heading, subtitle }) {
 
 	return (
 		<>
-			<LandingHero pageHeading={firstH1} pageSubtitle={firstP} />
+			<LandingHero pageHeading={heading} pageSubtitle={subtitle} />
 			{childrenToRender}
 		</>
 	)
