+++ src/views/product-root-docs-path-landing/components/hero/hero.module.css
@@ -1,29 +1,4 @@
 .root {
-  margin-bottom: 24px;
-  padding-bottom: 24px;
-}
-
-.iconAndTextWrapper {
-  align-items: center;
-  display: flex;
-  gap: 24px;
-  margin-bottom: 32px;
-}
-
-.icon {
-  display: none;
-
-  @media (--dev-dot-tablet-up) {
-    display: block;
-  }
-}
-
-.pageTitle {
-  composes: g-offset-scroll-margin-top from global;
-  color: var(--token-color-foreground-strong);
-  margin-bottom: 12px;
-}
-
-.pageSubtitle {
-  color: var(--token-color-foreground-primary);
+	margin-bottom: 24px;
+	padding-bottom: 24px;
 }
