+++ src/views/product-root-docs-path-landing/components/hero/index.tsx
@@ -1,7 +1,4 @@
-import { useCurrentProduct } from 'contexts'
-import IconTileLogo from 'components/icon-tile-logo'
-import Heading from 'components/heading'
-import Text from 'components/text'
+import LandingHero from 'components/landing-hero'
 import { ProductRootDocsPathLandingIconCardLinkGrid } from 'views/product-root-docs-path-landing/components'
 import { ProductRootDocsPathLandingHeroProps } from './types'
 import s from './hero.module.css'
@@ -10,30 +7,9 @@ const ProductRootDocsPathLandingHero = ({
 	pageHeading,
 	pageSubtitle,
 }: ProductRootDocsPathLandingHeroProps) => {
-	const currentProduct = useCurrentProduct()
-
 	return (
 		<div className={s.root}>
-			<div className={s.iconAndTextWrapper}>
-				<IconTileLogo
-					productSlug={
-						currentProduct.slug === 'sentinel' ? 'hcp' : currentProduct.slug
-					}
-					className={s.icon}
-				/>
-				<div>
-					<Heading
-						className={s.pageTitle}
-						id={pageHeading.id}
-						level={1}
-						size={500}
-						weight="bold"
-					>
-						{pageHeading.title}
-					</Heading>
-					<Text className={s.pageSubtitle}>{pageSubtitle}</Text>
-				</div>
-			</div>
+			<LandingHero pageHeading={pageHeading} pageSubtitle={pageSubtitle} />
 			<ProductRootDocsPathLandingIconCardLinkGrid />
 		</div>
 	)
